Update processing.gs
Added functionality to find the first lesson of the day. This will be used to set a pop up reminder in the calendar. This pop up will currently happen 10 hours before the lesson.

Some coding required to make this more dynamic based on the start time, morning, mid-morning or afternoon.