Add Summary Checker
Script to create references in the data sheet to make it easier to do
summary looks up.
Need to add an automation in scripts.google to run once a month, Though
only needs to run once on Jan 18th of any given year. If it is not
Janulary the script exists cleanly.